pyproject.toml
[project]
name = "askmebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "langchain>=0.3.26",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "supabase>=2.16.0",
    "uvicorn>=0.34.3",
    "python-dotenv>=1.0.0",
    "langchain-groq>=0.3.7",
]

[tool.uv.workspace]
members = [
    "fastapitut/fastapitut",
]
